Add files via upload
Write a program to monitor the flow of an item into and out of a warehouse.The warehouse will have numerous deliveries and shipments for the item (a widget) during the time period covered. A shipment out is billed at a profit of 50 percent over the cost of a widget. Unfortunately, each shipment received may have a different cost associated with it. The accountants for the firm have  instituted a first-in, first-out system for filling orders. This means that the oldest widgets are the first ones sent out to fill an order. This method of inventory can be represented using a queue. Each data record will consist of

S or O: Shipment received or an order to be sent
#: Quantity received or shipped out
Cost: Cost per widget (only for a shipment received)
Vendor: Character string that names company sent to or received from

Write the necessary functionalities to store the shipments received and to process orders. The output for an order will consist of the quantity and the total cost for all the widgets in the order. Hint: Each widget price is 50 percent higher than its cost. The widgets used to fill an order may come from multiple shipments with different costs.